+++ tests/unit_tests/plugin_tests/test_simulation_time_series/test_utils/test_derived_ensemble_vectors_accessor_utils.py
@@ -54,6 +54,7 @@ def test_create_derived_vectors_accessor_dict() -> None:
         expressions=[],
         delta_ensembles=delta_ensembles,
         resampling_frequency=None,
+        relative_date=None,
     )
 
     assert len(created_result) == len(ensembles)